Much work to Reference records
This sets up scopes for the Reference records. It also adds a new type
for the Reference records' options, in order to make the code more
readable and maintainable. This is a breaking change, as it changes the
structure of the Reference records. This change is necessary to make
the code more readable and maintainable.